5/27
I was at home and figured out how to move around the butter when you drag it, and how to “add it into” the bowl. I also added one more screen where customers are supposed to order what they want and an arrow that leads you to the cooking section.