6/11
I worked a lot to finally finish the game. I added one more scene to the game where there is a topping station. I also created images for each topping so it isn’t just a printed text that comes up. I also changed the name of the entire game and created a mascot, Bessie, that asks for an order.